+++ client/components/mma/cancel/digitalShared/SharedSubscriptionLeaveConfirmation.tsx
@@ -1,31 +1,20 @@
 import { css } from '@emotion/react';
 import {
 	from,
-	headlineBold24,
-	headlineBold28,
 	space,
 	textSans15,
 	textSans17,
 } from '@guardian/source/foundations';
-import { Button } from '@guardian/source/react-components';
+import { Button, Stack } from '@guardian/source/react-components';
 import { useNavigate } from 'react-router-dom';
 import { DATE_FNS_LONG_OUTPUT_FORMAT, dateString } from '@/shared/dates';
-import { NAV_LINKS } from '../../shared/nav/NavConfig';
-import { WithStandardTopMargin } from '../../shared/WithStandardTopMargin';
-
-const titleCss = css`
-	${headlineBold24};
-	margin: 0 0 ${space[4]}px;
-
-	${from.tablet} {
-		${headlineBold28};
-		margin-bottom: ${space[5]}px;
-	}
-`;
+import { NAV_LINKS } from '../../../shared/nav/NavConfig';
+import { Heading } from '../../shared/Heading';
+import { ctaContainerCss, titleCss } from '../cancellationConstants';
 
 const bodyCss = css`
 	${textSans15};
-	margin: 0 0 ${space[6]}px;
+	margin: ${space[2]}px 0;
 
 	${from.tablet} {
 		${textSans17};
@@ -39,24 +28,32 @@ export const LeaveSharedSubConfirmation = () => {
 	const accessEndDate = dateString(new Date(), DATE_FNS_LONG_OUTPUT_FORMAT);
 
 	return (
-		<WithStandardTopMargin>
-			<h2 css={titleCss}>You have left the digital plus subscription</h2>
+		<Stack space={3}>
+			<Heading borderless={true} cssOverrides={titleCss} level={'2'}>
+				You have left the digital plus subscription
+			</Heading>
+
 			<p css={bodyCss}>
 				You can continue to enjoy Digital plus benefits until{' '}
 				{accessEndDate}. After this date, you’ll no longer have access
 				to Digital plus benefits through this shared subscription.
-				<br />
+			</p>
+
+			<p css={bodyCss}>
 				If you have any questions, please contact our Customer service
 				team.
 			</p>
-			<Button
-				priority="primary"
-				onClick={() => {
-					navigate(NAV_LINKS.accountOverview.link);
-				}}
-			>
-				Return to account overview
-			</Button>
-		</WithStandardTopMargin>
+
+			<div css={ctaContainerCss}>
+				<Button
+					priority="primary"
+					onClick={() => {
+						navigate(NAV_LINKS.accountOverview.link);
+					}}
+				>
+					Return to account overview
+				</Button>
+			</div>
+		</Stack>
 	);
 };
