+++ client/components/mma/cancel/digitalShared/SharedSubscriptionLeaveContainer.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
-import { NAV_LINKS } from '../../shared/nav/NavConfig';
-import { PageContainer } from '../Page';
+import { NAV_LINKS } from '@/client/components/shared/nav/NavConfig';
+import { PageContainer } from '../../Page';
 
 export const LeaveSharedSubscriptionContainer = () => (
 	<PageContainer
