+++ client/components/mma/cancel/digitalShared/SharedSubscriptionLeaveFlowStart.tsx
@@ -1,66 +1,34 @@
-import { css } from '@emotion/react';
-import {
-	from,
-	headlineBold24,
-	headlineBold28,
-	space,
-	textSans15,
-	textSans17,
-} from '@guardian/source/foundations';
 import {
 	Button,
+	Stack,
 	SvgArrowLeftStraight,
 } from '@guardian/source/react-components';
 import { useNavigate } from 'react-router-dom';
-import { WithStandardTopMargin } from '../../shared/WithStandardTopMargin';
-
-const titleCss = css`
-	${headlineBold24};
-	margin: 0 0 ${space[4]}px;
-
-	${from.tablet} {
-		${headlineBold28};
-		margin-bottom: ${space[5]}px;
-	}
-`;
-
-const bodyCss = css`
-	${textSans15};
-	margin: 0 0 ${space[6]}px;
-
-	${from.tablet} {
-		${textSans17};
-	}
-`;
-
-const ctaContainerCss = css`
-	display: flex;
-	flex-direction: column;
-	gap: ${space[3]}px;
-
-	${from.tablet} {
-		flex-direction: row;
-	}
-`;
+import { Heading } from '../../shared/Heading';
+import { bodyCss, ctaContainerCss, titleCss } from '../cancellationConstants';
 
 export const LeaveSharedSubscription = () => {
 	const navigate = useNavigate();
 
 	return (
-		<WithStandardTopMargin>
-			<h2 css={titleCss}>
+		<Stack space={3}>
+			<Heading borderless={true} cssOverrides={titleCss} level={'2'}>
 				We’re sorry to see you go.
 				<br />
 				Are you sure you want to leave this subscription?
-			</h2>
+			</Heading>
+
 			<p css={bodyCss}>
 				If you leave this shared subscription, you’ll lose access to
 				Guardian premium benefits provided through it.
-				<br />
+			</p>
+			<p css={bodyCss}>
 				The person who invited you will be notified that you’ve left.
-				<br />
+			</p>
+			<p css={bodyCss}>
 				You’ll need a new invitation if you want to rejoin later.
 			</p>
+
 			<div css={ctaContainerCss}>
 				<Button
 					priority="tertiary"
@@ -72,7 +40,6 @@ export const LeaveSharedSubscription = () => {
 				>
 					Previous
 				</Button>
-				{/* TODO: implement leave shared subscription flow */}
 				<Button
 					priority="primary"
 					onClick={() => {
@@ -84,6 +51,6 @@ export const LeaveSharedSubscription = () => {
 					Yes, leave subscription
 				</Button>
 			</div>
-		</WithStandardTopMargin>
+		</Stack>
 	);
 };
